requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
djoser==2.2.3
mysqlclient==2.2.4
